move CLI code to /src/cli, and export in /dist/cli
