+++ packages/react-email/package.json
@@ -3,7 +3,7 @@
   "version": "5.3.0-canary.2",
   "description": "A live preview of your emails right in your browser.",
   "bin": {
-    "email": "./dist/index.js"
+    "email": "./dist/cli/index.js"
   },
   "type": "module",
   "scripts": {
