+++ packages/react-email/src/cli/commands/.npmignore
