+++ packages/react-email/src/cli/commands/resend/reset.ts
