+++ packages/react-email/src/cli/commands/resend/setup.ts
