+++ packages/react-email/src/cli/utils/esbuild/escape-string-for-regex.ts
