+++ packages/react-email/src/cli/utils/get-emails-directory-metadata.ts
