+++ packages/react-email/src/cli/utils/packageJson.ts
@@ -1,4 +1,4 @@
 // @ts-expect-error Typescript doesn't want to allow this, but it's fine since we're using tsup
-import packageJson from '../../package.json';
+import packageJson from '../../../package.json';
 
 export { packageJson };
