+++ packages/react-email/src/cli/utils/preview/serve-static-file.ts
