+++ packages/react-email/src/cli/utils/register-spinner-autostopping.ts
