+++ packages/react-email/src/cli/utils/style-text.ts
