+++ packages/react-email/src/cli/utils/types/hot-reload-event.ts
