+++ packages/react-email/tsdown.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from 'tsdown';
 
 export default defineConfig({
   dts: false,
-  entry: ['./src/index.ts'],
+  entry: ['./src/cli/index.ts'],
   format: ['esm'],
-  outDir: 'dist',
+  outDir: 'dist/cli',
 });
